refactor: Extract service interfaces to separate files
- Created Interfaces/IAuthService.cs and Interfaces/IPasswordService.cs
- Moved interface definitions from AuthService and PasswordService to respective interface files
- Updated namespace references to use new interface locations
